pyproject.toml
[project]
name = "abac-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-adk>=1.8.0",
    "google-cloud-modelarmor==0.2.6",
]
